+++ src/main/java/pro/sky/animal_shelter_telegram_bot/service/impl/PetServiceImpl.java
@@ -24,23 +24,27 @@ public class PetServiceImpl implements PetService {
     @Override
     public Pet addPet(Pet pet) {
         Pet addingPet = petRepository.save(pet);
-        logger.info("Pet {} is saved", addingPet);
+        logger.info("Pet with id {} is saved", addingPet);
         return addingPet;
     }
 
-    @Override
-    public void deletePet(Pet pet) {
-        petRepository.deleteById(pet.getId());
-        logger.info("Pet {} is deleted", pet);
-    }
-
-    public void deletePet(Long id) {
+    public Pet deletePet(Long id) {
+        if (petRepository.findById(id).isEmpty()){
+            logger.info("Pet with id {} is not found", id);
+            return null;
+        }
+        Pet deletePet = petRepository.findById(id).get();
         petRepository.deleteById(id);
         logger.info("Pet with id {} is deleted", id);
+        return deletePet;
     }
 
     @Override
     public Pet findPet(Long id) {
+        if (petRepository.findById(id).isEmpty()){
+            logger.info("Pet with id {} is not found", id);
+            return null;
+        }
         Pet findingPet = petRepository.findById(id).get();
         logger.info("Pet with id {} is found", id);
         return findingPet;
@@ -48,8 +52,12 @@ public class PetServiceImpl implements PetService {
 
     @Override
     public Pet changePet(Pet pet) {
+        if (petRepository.findById(pet.getId()).isEmpty()){
+            logger.info("Pet with id {} is not found", pet.getId());
+            return null;
+        }
         Pet changingPet = petRepository.save(pet);
-        logger.info("Pet {} is saved", pet);
+        logger.info("Pet with id {} is saved", pet);
         return changingPet;
     }
 }
