+++ src/main/java/pro/sky/animal_shelter_telegram_bot/service/impl/VolunteerServiceImpl.java
@@ -3,9 +3,13 @@ package pro.sky.animal_shelter_telegram_bot.service.impl;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
+import org.webjars.NotFoundException;
 import pro.sky.animal_shelter_telegram_bot.model.Volunteer;
 import pro.sky.animal_shelter_telegram_bot.repository.VolunteerRepository;
+import pro.sky.animal_shelter_telegram_bot.service.PetOwnerService;
 import pro.sky.animal_shelter_telegram_bot.service.VolunteerService;
+
+import java.util.List;
 
 /**
  * Service for working with repository VolunteerRepository
@@ -17,8 +21,11 @@ public class VolunteerServiceImpl implements VolunteerService {
 
     private final VolunteerRepository volunteerRepository;
 
-    public VolunteerServiceImpl(VolunteerRepository volunteerRepository) {
+    private final PetOwnerService petOwnerService;
+
+    public VolunteerServiceImpl(VolunteerRepository volunteerRepository, PetOwnerServiceImpl petOwnerService) {
         this.volunteerRepository = volunteerRepository;
+        this.petOwnerService = petOwnerService;
     }
 
     @Override
@@ -29,51 +36,86 @@ public class VolunteerServiceImpl implements VolunteerService {
     }
 
     @Override
-    public void deleteVolunteer(Volunteer volunteer) {
-        volunteerRepository.deleteById(volunteer.getId());
-        logger.info("Volunteer {} is deleted", volunteer);
-    }
-
-    @Override
-    public void deleteVolunteer(Long id) {
+    public Volunteer deleteVolunteer(Long id) {
+        if (volunteerRepository.findById(id).isEmpty()) {
+            logger.info("Volunteer with id {} is not found", id);
+            return null;
+        }
+        Volunteer deleteVolunteer = volunteerRepository.findById(id).get();
         volunteerRepository.deleteById(id);
         logger.info("Volunteer with id {} is deleted", id);
-
+        return deleteVolunteer;
     }
 
     @Override
     public Volunteer findVolunteer(Long id) {
-        Volunteer findingVolunteer = volunteerRepository.findById(id).get();
+        if (volunteerRepository.findById(id).isEmpty()) {
+            logger.info("Volunteer with id {} is not found", id);
+            return null;
+        }
+        Volunteer volunteer = volunteerRepository.findById(id).get();
         logger.info("Volunteer with id {} is found", id);
-        return findingVolunteer;
+        return volunteer;
     }
+
 
     @Override
     public Volunteer changeVolunteer(Volunteer volunteer) {
+        if (volunteerRepository.findById(volunteer.getId()).isEmpty()) {
+            logger.info("Volunteer with id {} is not found", volunteer.getId());
+            return null;
+        }
         Volunteer changingVolunteer = volunteerRepository.save(volunteer);
-        logger.info("Volunteer {} is saved", volunteer);
+        logger.info("Volunteer with id {} is saved", volunteer);
         return changingVolunteer;
     }
 
+
+
     /**
-     * add phone number to database
+     * add phone number to database. If this phone number was in pet-owner table, chat id will be saved to this table
      *
      * @param phoneNumber - phone number from swagger
-     * @param id          - volunteer id, positive
-     * @return
-     */
+     * @param volunteer   - volunteer
+       */
     @Override
-    public Volunteer setVolunteersPhoneNumber(String phoneNumber, Long id) {
+    public Volunteer setVolunteersPhoneNumber(Volunteer volunteer, String phoneNumber) {
         if (phoneNumber.isEmpty()) {
             logger.info("Phone number is empty");
             throw new NullPointerException("Phone number is empty");
         }
-        Volunteer changingVolunteer = volunteerRepository.findById(id).orElse(new Volunteer());
-        if (changingVolunteer == null) {
-            logger.info("Volunteer was not found, created empty Volunteer");
+        volunteer.setPhoneNumber(phoneNumber);
+        try {
+            volunteer.setChatId(petOwnerService.getPetOwnerChatIdByPhoneNumber(phoneNumber));
+        } catch (NullPointerException e) {
+            logger.info("Error");
         }
-        changingVolunteer.setPhoneNumber(phoneNumber);
-        logger.info("Volunteer {} is changed. Phone number {} is added.", changingVolunteer + phoneNumber);
-        return volunteerRepository.save(changingVolunteer);
+
+        logger.info("Volunteer {} is changed. Phone number {} is added.", volunteer, phoneNumber);
+        return volunteerRepository.save(volunteer);
+    }
+
+    /**
+     * find volunteer in database
+     * @param phoneNumber
+     * @return volunteer
+     */
+    @Override
+    public Volunteer findVolunteerByPhoneNumber(String phoneNumber) {
+        return volunteerRepository.findVolunteerByPhoneNumber(phoneNumber);
+    }
+
+    /**
+     *
+     * @return list of volunteers
+     */
+    @Override
+    public List<Volunteer> findAllVolunteer() {
+        List<Volunteer> volunteerList = volunteerRepository.findAll();
+        if (volunteerList.isEmpty()) {
+            logger.error("Volunteer list is empty");
+            throw new NotFoundException("Volunteer are empty");
+        }
+        return volunteerList;
     }
 }
